common: import toDbModel and implement for LobstersPostDetails
